+++ src/components/FSettingsPanel.ts
@@ -101,7 +101,7 @@ export class FSettingsPanel extends LitElement {
    * tests and the panel can read the same value without needing
    * a build-time injection mechanism.
    */
-  public static readonly VERSION = '2.0.0-beta.4';
+  public static readonly VERSION = '2.0.0-beta.5';
 
   @property({ type: Boolean })
   open = false;
@@ -259,9 +259,7 @@ export class FSettingsPanel extends LitElement {
                 ms
               </label>
             </fieldset>
-          </div>
 
-          <div class="fTelnetSettingsPanelColumn">
             <fieldset
               class="fTelnetSettingsPanelGroup fTelnetSettingsPanelGroupReserved"
             ></fieldset>